fixed type conversions

+++ demo/main.c
@@ -18,13 +18,13 @@ void fox_launch(struct FoxServer *server){
     char *hello = "HTTP/1.1 200 OK\nGMT\nServer: Apache/2.2.14 (Win32)\nLast-Modified: Wed, 22 Jul 2009 19:15:56 GMT\nContent-Type: text/html\nConnection: Closed\n\n<html><body><h1>Hello, Fox</h1></body></html>";
 
     int address_length = sizeof(server->address);
-    int new_socket = accept(server->socket, (struct  sockaddr*)&server->address, (int*)&address_length);
+    unsigned int new_socket = accept(server->socket, (struct  sockaddr*)&server->address, (int*)&address_length);
 
     while(1){
         printf("====== Waiting for connection ======");
         recv(new_socket,buffer,30000,0);
         printf("%s\n",buffer);
-        send(new_socket, hello, strlen(hello), 0 );
+        send(new_socket, hello, (int)strlen(hello), 0 );
         closesocket(new_socket);
     }
